--- frontend/src/global/backend/client.ts
@@ -1,2 +1,0 @@
-// 미정: swagger 이용한 OpenAPI 기반 자동 타입 생성 도입
-export const API_BASE_URL = process.env.NEXT_PUBLIC_API_BASE_URL || "";
