--- frontend/src/global/types/rsData.ts
@@ -1,5 +1,0 @@
-export type RsData<T> = {
-  code: number;
-  message: string;
-  data: T;
-};
